Fixed TextField::SetText
SetText assigns the new text content to row 0, which may not even exist
when the call is made. Now the call is forwarded to the internal Text
object's setContent method.

I also fixed Text::setContent to reset the caret's poisition, which was
not updated and it may be set to an invalid row or/and column.